requirements.txt
starlette
sentry-asgi
uvicorn
click
aioredis
aiocache
asyncpg
tortoise-orm
ipython
pyyaml
jinja2
webargs
